Replaced ints with proper types
I want to get in the habit of doing this so may aswell start now

+++ include/network/NetworkResponse.h
@@ -7,7 +7,7 @@
 class NetworkResponse
 {
     public:
-        int HTTPCode = 0;
+        unsigned short HTTPCode = 0;
         CURLcode CurlCode = CURLcode::CURLE_OK;
         std::string Body = "";
 };
