+++ src/platforms/SpotifyDownloader.cpp
@@ -21,8 +21,8 @@ ELinkType SpotifyDownloader::GetLinkType(const std::string& url)
 
 std::string SpotifyDownloader::GetLinkId(const std::string& url)
 {
-    int startIndex = url.find_last_of('/') + 1;
-    int endIndex = url.find('?');
+    unsigned short startIndex = url.find_last_of('/') + 1;
+    unsigned short endIndex = url.find('?');
     return url.substr(startIndex, endIndex - startIndex);
 }
